Add API controller Adjust routes
