+++ src/Volvox.Helios.Web/Controllers/SettingsController.cs
@@ -14,7 +14,7 @@ using Volvox.Helios.Web.ViewModels.Settings;
 namespace Volvox.Helios.Web.Controllers
 {
     [Authorize]
-    [Route("/Settings")]
+    [Route("/Settings/{guildId}")]
     public class SettingsController : Controller
     {
         private readonly IModuleSettingsService<StreamAnnouncerSettings> _streamAnnouncerSettingsService;
@@ -23,36 +23,16 @@ namespace Volvox.Helios.Web.Controllers
         {
             _streamAnnouncerSettingsService = streamAnnouncerSettingsService;
         }
-
-        [Route("{guildId}")]
+        
         public IActionResult Index(ulong guildId)
         {
             return View(guildId);
         }
 
-        // GET
-        [HttpGet("GetGuildChannels")]
-        public async Task<JsonResult> GetGuildChannels([FromServices] IDiscordGuildService guildService, ulong guildId)
-        {
-            var channels = await guildService.GetChannels(guildId);
-
-            // Format the ulong to string.
-            return Json(channels.FilterChannelType(0).Select(c => new {id = c.Id.ToString(), name = c.Name}));
-        }
-        
-        [HttpGet("GetUserAdminGuilds")]
-        public async Task<JsonResult> GetUserAdminGuilds([FromServices] IDiscordUserService userService)
-        {
-            var guilds = await userService.GetUserGuilds();
-
-            // Format the ulong to string.
-            return Json(guilds.FilterAdministrator());
-        }
-
         #region StreamAnnouncer
 
         // GET
-        [HttpGet("{guildId}/StreamAnnouncer")]
+        [HttpGet("StreamAnnouncer")]
         public async Task<IActionResult> StreamAnnouncerSettings(ulong guildId, [FromServices] IDiscordUserService userService,
             [FromServices] IBot bot)
         {
@@ -70,7 +50,7 @@ namespace Volvox.Helios.Web.Controllers
         }
 
         // POST
-        [HttpPost("{guildId}/StreamAnnouncer")]
+        [HttpPost("StreamAnnouncer")]
         public async Task<IActionResult> StreamAnnouncerSettings(StreamAnnouncerSettingsViewModel viewModel)
         {
             // Save the settings to the database
